# Conflicts:
#	Docker-compose/docker-compose.yml
#	Docker-compose/run-docker-compose.sh
#	iam-proxy-italia-project/static/disco.html

+++ Docker-compose/run-docker-compose.sh
@@ -1,28 +1,59 @@
 #!/bin/bash
 export COMPOSE_PROFILES=demo
+export SATOSA_CLEAN_DATA="false"
 export SKIP_UPDATE=
 export RUN_SPID_TEST=
 
 function clean_data {
-  rm -Rf ./mongo/db/*
-  rm -Rf ./iam-proxy-italia-project/*
-  rm -Rf ./djangosaml2_sp/*
-  rm -Rf ./nginx/html/static
+  if [ $SATOSA_CLEAN_DATA == "true" ]; then
+    rm -Rf ./mongo/db/*
+    rm -Rf ./iam-proxy-italia-project/*
+    rm -Rf ./djangosaml2_sp/*
+    rm -Rf ./nginx/html/static
+    rm -Rf ./certbot/live/localhost/*
+    rm -Rf ./spid_cie_oidc_django/wallet_trust_anchor/*
+    if [ "$SATOSA_FORCE_ENV" == "true" ]; then rm .env; fi
+  else
+    if [ "$SATOSA_FORCE_ENV" == "true" ]; then echo "'-e' options is skipped. To perform this option is required '-f' too "; fi
+  fi
+}
+
+function init_files () {
+  if [ -f $1 ]; then echo "$2 file is already initialized" ; else $3 ; fi
+}
+
+function add_localhost_cert () {
+  openssl req -x509 -out ./certbot/live/localhost/fullchain.pem -keyout certbot/live/localhost/privkey.pem \
+  -newkey rsa:2048 -nodes -sha256 \
+  -subj '/CN=localhost' -extensions EXT -config <( \
+   printf "[dn]\nCN=localhost\n[req]\ndistinguished_name = dn\n[EXT]\nsubjectAltName=DNS:localhost\nkeyUsage=digitalSignature\nextendedKeyUsage=serverAuth")
+}
+
+function add_iam_cert () {
+  cd ./iam-proxy-italia-project/pki
+  bash build_spid_certs.sh
+  cd ../..
 }
 
 function initialize_satosa {
-  cp env.example .env
-
   echo "WARNING: creating directories with read/write/execute permissions to anybody"
   
   mkdir -p ./iam-proxy-italia-project
   mkdir -p ./djangosaml2_sp
   mkdir -p ./mongo/db
   mkdir -p ./nginx/html/static
+  mkdir -p ./certbot/live/localhost
+  mkdir -p ./spid_cie_oidc_django/wallet_trust_anchor
 
-  if [ ! -f ./iam-proxy-italia-project/proxy_conf.yaml ]; then cp -R ../iam-proxy-italia-project/* ./iam-proxy-italia-project/ && rm -R ./satosa/static/ ; else echo 'iam-proxy-italia-project directory is already initialized' ; fi
-  if [ ! -f ./djangosaml2_sp/run.sh ]; then cp -R ../iam-proxy-italia-project_sp/djangosaml2_sp/* ./djangosaml2_sp ; else echo 'djangosaml2_sp directory is already initialided' ; fi
-  if [ ! -f ./nginx/html/static/disco.html ]; then cp -R ../iam-proxy-italia-project/static/* ./nginx/html/static ; else echo 'nginx directory is already initialized' ; fi
+  init_files ./.env ".env" "cp env.example .env"
+  init_files ./iam-proxy-italia-project/proxy_conf.yaml "iam-proxy-italia" "cp -R ../iam-proxy-italia-project ./"
+  init_files ./djangosaml2_sp/run.sh "djangosaml2_sp" "cp -R ../iam-proxy-italia-project-demo-examples/djangosaml2_sp ./"
+  init_files ./nginx/html/static/disco.html "static pages" "cp -R ../iam-proxy-italia-project/static ./nginx/html"
+  init_files ./certbot/live/localhost/privkey.pem "Locahost cert" "add_localhost_cert"
+  init_files ./iam-proxy-italia-project/pki/privkey.pem "IAM Proxy cert" "add_iam_cert"
+  init_files ./spid_cie_oidc_django/wallet_trust_anchor/manage.py "Wallet Trust Anchor" "cp -R ../iam-proxy-italia-project-demo-examples/spid_cie_oidc_django/wallet_trust_anchor ./spid_cie_oidc_django/"
+
+  rm -Rf ./iam-proxy-italia-project/static
 
   chmod -R 777 ./iam-proxy-italia-project
   echo "WARNING: iam-proxy-italia-project permission folder set recursively to 777"
@@ -37,12 +68,16 @@ function update {
     docker compose -f docker-compose.yml down -v
     echo -e "\n"
     echo -e "Tiro su la composizione, in caso, con le nuove versioni delle immagini. \n"
-    docker compose -f docker-compose.yml build django_sp
+    # docker compose -f docker-compose.yml build django_sp
   fi
 }
 
 function start {
-  docker compose -f docker-compose.yml up --wait --wait-timeout 60 --remove-orphans
+  if [ "$SATOSA_BUILD" == "true" ]; then
+    docker compose -f docker-compose.yml up --wait --wait-timeout 60 --remove-orphans --build
+  else
+    docker compose -f docker-compose.yml up --wait --wait-timeout 60 --remove-orphans
+  fi
   echo -e "\n"
   echo -e "Completato. Per visionare i logs: 'docker-compose -f docker-compose.yml logs -f'"
 
@@ -66,27 +101,39 @@ function start {
 
 function help {
   echo ""
-  echo "### run-docker-compose.sh ###"
+  echo "### run-docker-compose.sh"
   echo ""
   echo "initialize check update and start iam-proxy-italia compose structure"
   echo ""
-  echo "Options"
+  echo "#### common Options"
+  echo "-b Build for iam-proxy-italia image and build django-sp image if required"
+  echo "-e Force update for .env file. A new .env file is generated from env.example file. Require '-f' option otherwise is skipped"
   echo "-f Force clean and reinitialize data for Satosa, MongoDB and Djangosaml2_SP"
   echo "-h Print this help"
-  echo "-s Skip docker image update"
-  echo "-p unset compose profile. Run: satosa and nginx. Usefull for production"
+  echo ""
+  echo "#### profile options"
   echo "-m Set 'mongo' compose profile. Run: satosa, nginx, mongo"
   echo "-M Set 'mongoexpress' compose profile. Run: satosa, nginx, mongo, mongo-express"
+  echo "-p unset compose profile. Run: satosa and nginx. Usefull for production"
+  echo "-s Skip docker image update"
   echo "-d Set 'dev' compose profile. Run: satosa, nginx, django-sp, spid-saml-check"
   echo "-t Run spid_sp_test tests after startup"
-  echo "   if isn't set any of -p, -m, -M, -d, is used 'demo' compose profile"
-  echo "   demo compose profile start: satosa, nginx, mongo, mongo-express, django-sp, spid-saml-check"
+  echo ""
+  echo "if isn't set any options of -p, -m, -M, -d, is used 'demo' compose profile"
+  echo "demo compose profile start: satosa, nginx, mongo, mongo-express, django-sp, spid-saml-check"
+  echo ""
 }
 
-while getopts ":fpmMdsth" opt; do
+while getopts ":fepbimMdsh" opt; do
   case ${opt} in
    f)
-     clean_data
+     SATOSA_CLEAN_DATA="true"
+     ;;
+   e)
+     SATOSA_FORCE_ENV="true"
+     ;;
+   b)
+     SATOSA_BUILD="true"
      ;;
    p)
      unset COMPOSE_PROFILES
@@ -118,7 +165,7 @@ while getopts ":fpmMdsth" opt; do
      ;;
   esac
 done
-initialize_satosa
-update
-start
-
+clean_data         # clean docker compose directories if $SATOSA_CLEAN_DATA == "true"
+initialize_satosa  # check and initialize docker compose directories
+update             # try to update the images unless $SKIP_UPDATE is present
+start              # run docker compose
